requirements.txt
appdirs==1.4.4
asgiref==3.2.10
astroid==1.6.6
autopep8==1.5.1
backcall==0.2.0
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.2
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.1
cryptography==2.9.2
decorator==4.4.2
distlib==0.3.0
dj-database-url==0.5.0
Django==3.0
django-cors-headers==3.4.0
django-heroku==0.3.1
django-rest-swagger==2.2.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
docutils==0.16
drf-yasg==1.17.1
filelock==3.0.12
gunicorn==20.0.4
idna==2.9
inflection==0.5.0
ipdb==0.10.1
ipython==7.15.0
ipython-genutils==0.2.0
isort==4.3.21
itypes==1.2.0
jedi==0.17.1
Jinja2==2.11.2
lazy-object-proxy==1.4.3
livereload==2.6.2
Markdown==3.2.2
MarkupSafe==1.1.1
mccabe==0.6.1
mkdocs==0.15.3
mkdocs-bootstrap==1.1
mkdocs-bootswatch==1.1
openapi-codec==1.3.2
packaging==20.4
parso==0.7.0
pickleshare==0.7.5
pluggy==0.4.0
prompt-toolkit==3.0.5
psycopg2==2.8.5
py==1.8.2
pycodestyle==2.5.0
pycparser==2.20
Pygments==2.6.1
PyJWT==1.7.1
pylint==1.8.4
pyOpenSSL==19.1.0
pyparsing==2.4.7
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
simplejson==3.9.0
six==1.14.0
sqlparse==0.3.1
tornado==6.0.4
tox==2.4.1
traitlets==4.3.3
Unidecode==1.1.1
uritemplate==3.0.1
urllib3==1.25.9
virtualenv==20.0.25
wcwidth==0.2.5
whitenoise==5.1.0
wrapt==1.11.2
